% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Weighted Linear Calibration from instrumental data}
\usage{
calibrate(signal, RSD, conc, element, conc_unit, signal_unit)
}
\arguments{
\item{signal}{attribute}

\item{RSD}{attribute}

\item{conc}{attribute}

\item{element}{character}

\item{conc_unit}{character}

\item{signal_unit}{character}
}
\value{

}
\description{
Weighted Linear Calibration from instrumental data
}
\examples{
calibrate(ICPMS$signal, ICPMS$RSD, ICPMS$conc, "Pb", "um", "ppb")

}
